# openapi.yaml
openapi: 3.0.3
info:
  title: Fleet Management API
  description: API for managing fleet vehicles, drivers, maintenance records, and assignments.
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

# No authentication required (optional; omit or add security schemes later)
security: []
servers:
  - url: http://localhost:5298/api
    description: Local API
  - url: /api
    description: Relative base path
...
components:
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: Human-readable error description
    VehicleStatus:
      type: integer
      enum: [0, 1, 2]
      description: 0 = Available, 1 = NotAvailable, 2 = UnderMaintenance
...
    MaintenanceType:
      type: integer
      enum: [0, 1, 2, 3, 4, 5, 6, 7]
      description: 0 = Cleaning, 1 = Washing, 2 = OilChange, 3 = TireReplacement, 4 = BrakeService, 5 = Inspection, 6 = Repair, 7 = Other
    AssignmentStatus:
      type: integer
      enum: [0, 1, 2]
      description: 0 = Active, 1 = Completed, 2 = Cancelled
    Vehicle:
      type: object
      description: Vehicle entity
      properties:
        id: { type: string, format: uuid, description: "Unique identifier (read-only, server-assigned)" }
        make: { type: string, description: "Vehicle manufacturer (e.g. Toyota, Ford)" }
        model: { type: string, description: Vehicle model name }
        year: { type: integer, description: Model year }
        vin: { type: string, description: 17-character Vehicle Identification Number }
        licensePlate: { type: string, description: License plate number }
        registrationDate: { type: string, format: date-time, description: Date of registration }
        purchasePrice: { type: number, description: Purchase price }
        currentValue: { type: number, description: Current estimated value }
        status: { $ref: '#/components/schemas/VehicleStatus' }
        createdAt: { type: string, format: date-time, description: "When the record was created (server-set)" }
        updatedAt: { type: string, format: date-time, description: "When the record was last updated (server-set)" }
        isDeleted: { type: boolean, default: false, description: "Soft-delete flag (server-managed)" }
    VehicleCreate:
      type: object
      description: Request body for creating a vehicle (omit id, createdAt, updatedAt, isDeleted)
      required: [make, model, year, vin, licensePlate, registrationDate, purchasePrice, currentValue, status]
      properties:
        make: { type: string }
        model: { type: string }
        year: { type: integer }
        vin: { type: string }
        licensePlate: { type: string }
        registrationDate: { type: string, format: date-time }
        purchasePrice: { type: number }
        currentValue: { type: number }
        status: { $ref: '#/components/schemas/VehicleStatus' }
...
    MaintenanceRecord:
      type: object
      description: Maintenance record entity
      properties:
        id: { type: string, format: uuid, description: "Unique identifier (read-only)" }
        vehicleId: { type: string, format: uuid, description: ID of the vehicle }
        maintenanceType: { $ref: '#/components/schemas/MaintenanceType' }
        description: { type: string, nullable: true, description: Optional description }
        cost: { type: number, description: Cost of the service }
        serviceDate: { type: string, format: date-time, description: Date of service }
        serviceProvider: { type: string, nullable: true, description: Optional provider name }
        nextServiceDueDate: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        isDeleted: { type: boolean }
    MaintenanceRecordCreate:
      type: object
      description: Request body for creating a maintenance record
      required: [vehicleId, maintenanceType, cost, serviceDate, nextServiceDueDate]
      properties:
        vehicleId: { type: string, format: uuid }
        maintenanceType: { $ref: '#/components/schemas/MaintenanceType' }
        description: { type: string, nullable: true }
        cost: { type: number }
        serviceDate: { type: string, format: date-time }
        serviceProvider: { type: string, nullable: true }
        nextServiceDueDate: { type: string, format: date-time }
    Assignment:
      type: object
      description: Assignment entity (vehicle-driver assignment)
      properties:
        id: { type: string, format: uuid, description: "Unique identifier (read-only)" }
        vehicleId: { type: string, format: uuid, description: ID of the assigned vehicle }
        driverId: { type: string, format: uuid, description: ID of the assigned driver }
        startDate: { type: string, format: date-time, description: Assignment start date }
        endDate: { type: string, format: date-time, description: Assignment end date }
        status: { $ref: '#/components/schemas/AssignmentStatus' }
        notes: { type: string, nullable: true, description: Optional notes }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        isDeleted: { type: boolean }
...
paths:
  /vehicles:
    get:
      tags: [Vehicles]
      summary: Get all vehicles
      description: Returns all non-deleted vehicles.
      operationId: getVehicles
      responses:
        '200':
          description: List of vehicles
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Vehicle' }
    post:
      tags: [Vehicles]
      summary: Create vehicle
      description: Creates a new vehicle.
      operationId: createVehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleCreate'
      responses:
        '201':
          description: Vehicle created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Vehicle' }
        '400':
          description: Validation error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /vehicles/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
        description: Vehicle ID
    get:
      tags: [Vehicles]
      summary: Get vehicle by ID
      description: Returns a single vehicle by ID.
      operationId: getVehicle
      responses:
        '200':
          description: Vehicle found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Vehicle' }
        '404':
          description: Vehicle not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    put:
      tags: [Vehicles]
      summary: Update vehicle
      description: Updates an existing vehicle. Body id must match path id.
      operationId: updateVehicle
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Vehicle' }
      responses:
        '204':
          description: Update successful
        '400':
          description: Vehicle ID mismatch
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Vehicle not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [Vehicles]
      summary: Delete vehicle
      description: Soft-deletes the vehicle (sets isDeleted = true).
      operationId: deleteVehicle
      responses:
        '204':
          description: Soft-delete successful
        '404':
          description: Vehicle not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /vehicles/{id}/maintenance-records:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
        description: Vehicle ID
    get:
      tags: [Vehicles]
      summary: Get vehicle maintenance records
      description: Returns all maintenance records for the given vehicle.
      operationId: getVehicleMaintenanceRecords
      responses:
        '200':
          description: List of maintenance records
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MaintenanceRecord' }
        '404':
          description: Vehicle not found or no records
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /vehicles/{id}/assignments:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
        description: Vehicle ID
    get:
      tags: [Vehicles]
      summary: Get vehicle assignments
      description: Returns all assignments for the given vehicle.
      operationId: getVehicleAssignments
      responses:
        '200':
          description: List of assignments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Assignment' }
        '404':
          description: Vehicle not found or no assignments
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /maintenancerecords:
    get:
      tags: [Maintenance Records]
      summary: Get all maintenance records
      description: Returns all non-deleted maintenance records.
      operationId: getMaintenanceRecords
      responses:
        '200':
          description: List of maintenance records
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MaintenanceRecord' }
    post:
      tags: [Maintenance Records]
      summary: Create maintenance record
      description: Creates a new maintenance record. vehicleId must reference an existing non-deleted vehicle.
      operationId: createMaintenanceRecord
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MaintenanceRecordCreate' }
      responses:
        '201':
          description: Maintenance record created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MaintenanceRecord' }
        '400':
          description: Vehicle not found or validation error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /maintenancerecords/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
        description: Maintenance record ID
    get:
      tags: [Maintenance Records]
      summary: Get maintenance record by ID
      description: Returns a single maintenance record by ID.
      operationId: getMaintenanceRecord
      responses:
        '200':
          description: Maintenance record found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MaintenanceRecord' }
        '404':
          description: Maintenance record not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    put:
      tags: [Maintenance Records]
      summary: Update maintenance record
      description: Updates an existing maintenance record. If vehicleId changes, new vehicle must exist.
      operationId: updateMaintenanceRecord
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MaintenanceRecord' }
      responses:
        '204':
          description: Update successful
        '400':
          description: ID mismatch or vehicle not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Maintenance record not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [Maintenance Records]
      summary: Delete maintenance record
      description: Soft-deletes the maintenance record.
      operationId: deleteMaintenanceRecord
      responses:
        '204':
          description: Soft-delete successful
        '404':
          description: Maintenance record not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...